pyproject.toml
[project]
name = "dcf-modeling-tool"
version = "1.0.0"
description = "Production-quality DCF modeling tool with FCFF/WACC and FCFE/Ke valuation methods"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "typer>=0.9.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "openpyxl>=3.1",
    "pandas>=2.0",
    "plotly>=5.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
dcf = "dcf_ui_cli.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
